moved app.js to /containers
