+++ src/index.js
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import rootReducer from './reducers'
 import './index.css';
-import App from './App';
+import App from './containers/App/App';
 import registerServiceWorker from './registerServiceWorker';
 
 const boundCompose = compose.bind(null, applyMiddleware(thunk));
@@ -14,5 +14,8 @@ const store = createStore(
   window.__REDUX_DEVTOOLS_EXTENSION__ ? boundCompose(window.__REDUX_DEVTOOLS_EXTENSION__()) : boundCompose()
 );
 
-ReactDOM.render(<App />, document.getElementById('root'));
-registerServiceWorker();
+ReactDOM.render(<Provider store={store}>
+  <App />
+  </Provider>,
+   document.getElementById('root'));
+  registerServiceWorker();
